IGNITE-19566 Rename ExceptionUtils to TraceIdGenerator
