+++ modules/api/src/main/java/org/apache/ignite/lang/IgniteCheckedException.java
@@ -22,7 +22,7 @@ import static org.apache.ignite.lang.ErrorGroup.errorGroupByCode;
 import static org.apache.ignite.lang.ErrorGroup.errorMessage;
 import static org.apache.ignite.lang.ErrorGroup.extractErrorCode;
 import static org.apache.ignite.lang.ErrorGroup.extractGroupCode;
-import static org.apache.ignite.lang.util.ExceptionUtils.getOrCreateTraceId;
+import static org.apache.ignite.lang.util.TraceIdGenerator.getOrCreateTraceId;
 
 import java.util.UUID;
